% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generico.R
\name{generico}
\alias{generico}
\title{generico}
\usage{
generico(tabela, gradiente, at, grad, eixoY, eixoX)
}
\arguments{
\item{tabela}{A data.frame with the species data. Species in columns.}

\item{gradiente}{Vector with the environmental gradient for each row of the species table.}

\item{at}{Used to alter position of the species names. Start with 1 and increase until obtain desired location.}

\item{grad}{Character of the name of the environmental gradient to appear on the graph.}

\item{eixoY}{Name of y axis legend (character)}

\item{eixoX}{Name of x axis legend (character)}
}
\value{
Plots bar charts of species abundances ordered along the gradient.
}
\description{
This fuction is used to make direct ordination of species abundances.
}
\details{
Ported from 
<https://ppbio.inpa.gov.br/sites/default/files/Generico%20R.doc>
Tried and tested code used to develop package and github.
}
\examples{
\dontrun{
generico()
}
}
\author{
Created by Victor Lemes Landeiro. Updated 18-07-2008.
}

