% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generico.R
\name{generico}
\alias{generico}
\title{generico}
\usage{
generico(tabela, gradiente, at, grad, eixoY, eixoX)
}
\arguments{
\item{tabela}{Species data (object class: \code{data.frame}). Species in columns.}

\item{gradiente}{Vector with the environmental gradient for each row of the species table.}

\item{at}{Used to alter plotted position of the species names. Start with 1 and increase until obtain desired location.}

\item{grad}{Character of the name of the environmental gradient to appear on the graph.}

\item{eixoY}{Name of y axis legend (character)}

\item{eixoX}{Name of x axis legend (character)}
}
\value{
Function \code{generico} plots bar charts of species abundances from sites 
ordered along an environmental gradient.
}
\description{
\code{generico} plots direct ordination of species abundances.
}
\details{
Ported from: \url{https://ppbio.inpa.gov.br/sites/default/files/Generico\%20R} . 
 Tried and tested code used to develop package and github. Plots the standardized relative abundance of species in samples
 ordered in relation to an environmental gradient. Species are plotted seperately. This ordering provides
 visual representation of the counts of different species that enables a clearly interpretable comparison of
 which species are found at the extremes of a gradient or occur through the range. 
 Most effective with relatively few species (i.e. < 30).
 
 The function is not general. It is not possible for users to ajust graphical parameters.
 This means that the graph presented is unlikely to be correct for publication.
 Function code is likely to need adaptation for specific use cases (e.g. more or less species).
 Used to teach students how to adapt code from existing functions.
}
\examples{
\dontrun{
# load species and environmental data
library(vegan) 
data(mite, mite.env)
dfsp <- mite[,1:10] #select first 10 mite species 

#plot 
generico(tabela = dfsp, gradiente = mite.env$WatrCont,
at=80,grad = "Water",eixoX="Site", eixoY="Count") 
}
}
\author{
Created by Victor Lemes Landeiro. Updated 18-07-2008.
}
\references{
Examples of use:
  
Norris, D. et al. 2016. Too rare for non-timber resource harvest? 
Meso-scale composition and distribution of arborescent palms in an Amazonian sustainable-use forest.
Forest Ecology and Management 377: 182-191. \url{http://dx.doi.org/10.1016/j.foreco.2016.07.008} .
}

